适配DirectX 11
为了适配DirectX 11，修改了gfx\FX\下pdxmap.shader及pdxwater.shader两个文件

+++ gfx/FX/pdxwater.shader
@@ -154,7 +154,7 @@ PixelShader =
 
 VertexStruct VS_INPUT_WATER
 {
-    float2 position			: POSITION;
+    int2 position			: POSITION;
 };
 
 VertexStruct VS_OUTPUT_WATER
